% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions_GC.R
\name{logl.gp.mh.st}
\alias{logl.gp.mh.st}
\title{Inverse-Gamma DLM MCMC Simulation}
\usage{
logl.gp.mh.st(i, y, locstime, parms, h, m)
}
\arguments{
\item{dat}{vector of length \emph{T} containing a time series of VI data}
}
\value{
List containing MCMC samples of \eqn{\sigma_e^2}, \eqn{\sigma_w^2}, and \eqn{\theta_{1:T}} (if \code{save.theta = TRUE}) 
\describe{
\item{sigma2s:}{mc x 2 matrix containing MCMC samples of \eqn{\sigma_e^2} (first column) and \eqn{\sigma_w^2} (second column)}
\item{thetas:}{if \code{save_value = TRUE}, a T x p x mc array containing MCMC samples of the latent states}
}
}
\description{
MCMC simulation of the reduced Fourier-form Bayesian DLM with conjugate Inverse-Gamma priors.
}
\details{
blah blah blah
}
